@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "nmdc-server"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.12"
 dynamic = ["version"]
 dependencies = [
     "alembic==1.13.3",
@@ -44,11 +44,12 @@ dependencies = [
     "executing==2.1.0",
     "factory-boy==3.3.1",
     "faker==30.3.0",
-    "fastapi==0.115.0",
+    "fastapi[standard]==0.115.0",
     "flexcache==0.3",
     "flexparser==0.3.1",
     "fqdn==1.5.1",
     "ghp-import==2.1.0",
+    "google-cloud-storage==3.1.1",
     "graphviz==0.20.3",
     "greenlet==3.1.1",
     "h11==0.14.0",
@@ -74,9 +75,15 @@ dependencies = [
     "jsonschema==4.23.0",
     "jsonschema-specifications==2024.10.1",
     "kombu==5.4.2",
-    "linkml==1.8.5",
+    # Notes:
+    # - Updated from `==1.8.5` to `==1.9.1` to be compatible with `nmdc-schema` version `11.7.0`.
+    # - Chose version `==1.9.1`, specifically, because that is what `nmdc-runtime` currently uses.
+    "linkml==1.9.1",
     "linkml-dataops==0.1.0",
-    "linkml-runtime==1.8.3",
+    # Notes:
+    # - Updated from `==1.8.3` to `==1.9.2` to be compatible with `linkml` version `1.9.1`.
+    # - Chose version `==1.9.2`, specifically, because that is what `nmdc-runtime` currently uses.
+    "linkml-runtime==1.9.2",
     "mako==1.3.5",
     "markdown==3.7",
     "markupsafe==3.0.1",
@@ -90,8 +97,8 @@ dependencies = [
     "mkdocs-redirects==1.2.1",
     "mypy==1.11.2",
     "mypy-extensions==1.0.0",
-    "nmdc-schema==11.5.1",
-    "nmdc-submission-schema==11.5.0",
+    "nmdc-schema==11.9.1",
+    "nmdc-submission-schema==11.10.0-rc.1",
     "nmdc-geoloc-tools==0.2.0",
     "openpyxl==3.1.5",
     "packaging==24.1",
