@@ -1,32 +1,9 @@
-import re
 from enum import Enum
 from typing import Optional
 
 from pydantic import BaseModel
 
 from nmdc_server import models
-
-# Nginx's mod_zip can only server local files.  To get arround this limitation,
-# we set up local proxies to all remote hosts.
-#   https://www.nginx.com/resources/wiki/modules/zip/#remote-upstreams
-# This means that we can only handle known prefixes.  This must be checked
-# on ingest and any additional hosts added to the nginx config.
-# TODO: There is probably a way to automate this using nginx pattern matching.
-data_url_hosts = [
-    (re.compile("^https://data.microbiomedata.org(/data)?"), "/data"),
-    (re.compile("^https://nmdcdemo.emsl.pnnl.gov"), "/nmdcdemo"),
-    (re.compile("^https://portal.nersc.gov"), "/nerscportal"),
-    (re.compile("^https://storage.neonscience.org"), "/neonscience"),
-]
-
-
-def get_local_data_url(url: Optional[str]) -> Optional[str]:
-    if not url:
-        return None
-    for r, v in data_url_hosts:
-        if r.match(url):
-            return r.sub(v, url)
-    return None
 
 
 class WorkflowActivityTypeEnum(Enum):
@@ -35,7 +12,7 @@ class WorkflowActivityTypeEnum(Enum):
     metagenome_assembly = "nmdc:MetagenomeAssembly"
     metagenome_annotation = "nmdc:MetagenomeAnnotation"
     metaproteomic_analysis = "nmdc:MetaproteomicsAnalysis"
-    metatranscriptome = "nmdc:MetatranscriptomeAnalysis"
+    metatranscriptome_expression = "nmdc:MetatranscriptomeExpressionAnalysis"
     metatranscriptome_assembly = "nmdc:MetatranscriptomeAssembly"
     metatranscriptome_annotation = "nmdc:MetatranscriptomeAnnotation"
     nom_analysis = "nmdc:NomAnalysis"
@@ -64,7 +41,7 @@ _workflow_enum_to_model = {
     WorkflowActivityTypeEnum.nom_analysis: models.NOMAnalysis,
     WorkflowActivityTypeEnum.metabolomics_analysis: models.MetabolomicsAnalysis,
     WorkflowActivityTypeEnum.raw_data: models.OmicsProcessing,
-    WorkflowActivityTypeEnum.metatranscriptome: models.Metatranscriptome,
+    WorkflowActivityTypeEnum.metatranscriptome_expression: models.Metatranscriptome,
 }
 
 _mpa = WorkflowActivityTypeEnum.metaproteomic_analysis
@@ -81,7 +58,7 @@ _workflow_enum_to_output_association = {
     WorkflowActivityTypeEnum.nom_analysis: models.nom_analysis_output_association,
     WorkflowActivityTypeEnum.metabolomics_analysis: models.metabolomics_analysis_output_association,
     WorkflowActivityTypeEnum.raw_data: models.omics_processing_output_association,
-    WorkflowActivityTypeEnum.metatranscriptome: models.metabolomics_analysis_output_association,
+    WorkflowActivityTypeEnum.metatranscriptome_expression: models.metabolomics_analysis_output_association,  # noqa: E501
 }
 
 
