@@ -7,7 +7,7 @@ import re
 from datetime import datetime
 from enum import Enum
 from itertools import groupby
-from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
+from typing import Any, Dict, Generic, Iterator, List, Optional, Tuple, TypeVar, Union
 
 from pydantic import BaseModel, ConfigDict, Field, PositiveInt
 from sqlalchemy import ARRAY, Column, and_, cast, desc, func, inspect, or_
@@ -17,7 +17,7 @@ from sqlalchemy.sql.expression import ClauseElement, intersect, union
 from sqlalchemy.sql.selectable import CTE
 from typing_extensions import Annotated, Literal
 
-from nmdc_server import binning, models, schemas, schemas_submission
+from nmdc_server import binning, models, schemas
 from nmdc_server.binning import DateBinResolution
 from nmdc_server.data_object_filters import DataObjectFilter
 from nmdc_server.filters import create_filter_class
@@ -92,6 +92,10 @@ _envo_keys: Dict[str, Tuple[Table, str]] = {
 
 _association_proxy_keys: Dict[str, Tuple[Any, Any]] = {
     "principal_investigator_name": (models.Study, models.PrincipalInvestigator.name),
+    "metaproteomics_analysis_category": (
+        models.OmicsProcessing,
+        models.MetaproteomicAnalysis.metaproteomics_analysis_category,
+    ),
 }
 
 _special_keys: Dict[str, Tuple[Table, str]] = {
@@ -607,9 +611,10 @@ class StudyQuerySchema(BaseQuerySchema):
     # The following private methods inject subqueries into Study `query_expressions`.
     def _count_omics_data_query(self, db: Session, query_schema: BaseQuerySchema) -> Query:
         """Generate a query counting matching omics_processing types."""
-        model = query_schema.table.model
-        aliased_model = aliased(model)
-        table_name = model.__tablename__  # type: ignore
+        workflow_model = query_schema.table.model
+        aliased_workflow_model = aliased(workflow_model)
+        table_name = workflow_model.__tablename__  # type: ignore
+        was_informed_by_table = models.workflow_activity_to_data_generation_map[table_name]
 
         op_alias = aliased(models.OmicsProcessing)
         biosample_alias = aliased(models.Biosample)
@@ -619,12 +624,16 @@ class StudyQuerySchema(BaseQuerySchema):
         q = (
             db.query(
                 op_alias.study_id.label(f"{table_name}_study_id"),
-                func.count(func.distinct(model.id)).label(f"{table_name}_count"),  # type: ignore
+                func.count(func.distinct(aliased_workflow_model.id)).label(f"{table_name}_count"),
             )
-            .join(op_alias)
+            .select_from(op_alias)
             .join(biosample_alias, op_alias.biosample_inputs)
-            .join(aliased_model)
-            .join(subquery, subquery.c.id == model.id)  # type: ignore
+            .join(was_informed_by_table, was_informed_by_table.c.data_generation_id == op_alias.id)
+            .join(
+                aliased_workflow_model,
+                aliased_workflow_model.id == was_informed_by_table.c[f"{table_name}_id"],
+            )
+            .join(subquery, subquery.c.id == aliased_workflow_model.id)
             .group_by(op_alias.study_id)
         )
         return q
@@ -696,7 +705,8 @@ class StudyQuerySchema(BaseQuerySchema):
         op_filter_conditions = [
             c
             for c in self.conditions
-            if c.table.value in {"omics_processing", "biosample", "gene_function"}
+            if c.table.value
+            in {"omics_processing", "biosample", "gene_function", "metaproteomic_analysis"}
         ]
         op_summary_subquery = self._count_omics_processing_summary(
             db, op_filter_conditions
@@ -768,6 +778,34 @@ class OmicsProcessingQuerySchema(BaseQuerySchema):
     @property
     def table(self) -> Table:
         return Table.omics_processing
+
+    def facet(self, db: Session, attribute: str) -> Dict[schemas.AnnotationValue, int]:
+        if attribute == "omics_type":
+            annotations_column = models.OmicsProcessing.annotations
+            id_column = models.OmicsProcessing.poolable_replicates_manifest_id
+            query = self.query(db)
+            aggregated_query = query.with_entities(
+                func.count(func.distinct(id_column)).label("id_count"),
+                annotations_column["omics_type"].astext,
+            ).group_by(annotations_column["omics_type"].astext)
+            return {value: count for count, value in aggregated_query if value is not None}
+        if attribute == "metaproteomics_analysis_category":
+            association_table = models.metaproteomic_analysis_data_generation_association
+            query = self.query(db)
+            query = query.join(
+                association_table,
+                models.OmicsProcessing.id == association_table.c.data_generation_id,
+            )
+            query = query.join(
+                models.MetaproteomicAnalysis,
+                models.MetaproteomicAnalysis.id == association_table.c.metaproteomic_analysis_id,
+            )
+            aggregated_query = query.with_entities(
+                func.count(func.distinct(models.OmicsProcessing.id)).label("id_count"),
+                models.MetaproteomicAnalysis.metaproteomics_analysis_category,
+            ).group_by(models.MetaproteomicAnalysis.metaproteomics_analysis_category)
+            return {value: count for count, value in aggregated_query if value is not None}
+        return super().facet(db, attribute)
 
     def omics_processing_for_biosample_ids(self, db: Session, biosample_ids):
         # Do the normal query with the conditions
@@ -940,9 +978,16 @@ class DataObjectQuerySchema(BaseQuerySchema):
         self, db: Session, filter: DataObjectFilter, op_cte: CTE
     ) -> Query:
         """Create a subquery that selects from a data object filter condition."""
-        query = db.query(models.DataObject.id.label("id")).join(
-            op_cte,
-            models.DataObject.omics_processing_id == op_cte.c.id,
+        query = (
+            db.query(models.DataObject.id.label("id"))
+            .join(
+                models.omics_processing_output_association,
+                models.omics_processing_output_association.c.data_object_id == models.DataObject.id,
+            )
+            .join(
+                op_cte,
+                models.omics_processing_output_association.c.omics_processing_id == op_cte.c.id,
+            )
         )
         if filter.workflow:
             query = query.filter(models.DataObject.workflow_type == filter.workflow.value)
@@ -978,8 +1023,12 @@ class BaseSearchResponse(BaseModel):
     count: int
 
 
-class BiosampleSearchResponse(BaseSearchResponse):
-    results: List[schemas.Biosample]
+T = TypeVar("T")
+
+
+class Paginated(BaseModel, Generic[T]):
+    count: int
+    results: List[T]
 
 
 class SearchQuery(BaseModel):
@@ -1013,17 +1062,8 @@ class BinnedDateFacetQuery(FacetQuery):
 BinnedFacetQuery = Union[BinnedRangeFacetQuery, BinnedDateFacetQuery]
 
 
-class StudySearchResponse(BaseSearchResponse):
-    results: List[schemas.Study]
+class StudySearchResponse(Paginated[schemas.Study]):
     total: Optional[int] = None
-
-
-class OmicsProcessingSearchResponse(BaseSearchResponse):
-    results: List[schemas.OmicsProcessing]
-
-
-class DataObjectSearchResponse(BaseSearchResponse):
-    results: List[schemas.DataObject]
 
 
 class FacetResponse(BaseModel):
@@ -1033,10 +1073,6 @@ class FacetResponse(BaseModel):
 class BinnedFacetResponse(BaseModel):
     facets: List[int]
     bins: List[NumericValue]
-
-
-class MetadataSubmissionResponse(BaseSearchResponse):
-    results: List[schemas_submission.SubmissionMetadataSchema]
 
 
 class UserResponse(BaseSearchResponse):
