@@ -367,37 +367,37 @@ def test_pipeline_query(db: Session, table):
     omics_processing1 = fakes.OmicsProcessingFactory(name="omics_processing1")
     omics_processing2 = fakes.OmicsProcessingFactory(name="omics_processing2")
 
-    fakes.ReadsQCFactory(omics_processing=omics_processing1, name="reads_qc1")
-    fakes.ReadsQCFactory(omics_processing=omics_processing1, name="reads_qc2")
-    fakes.ReadsQCFactory(omics_processing=omics_processing2, name="reads_qc3")
+    fakes.ReadsQCFactory(was_informed_by=[omics_processing1], name="reads_qc1")
+    fakes.ReadsQCFactory(was_informed_by=[omics_processing1], name="reads_qc2")
+    fakes.ReadsQCFactory(was_informed_by=[omics_processing2], name="reads_qc3")
 
-    fakes.MetagenomeAssemblyFactory(omics_processing=omics_processing1, name="assembly1")
-    fakes.MetagenomeAssemblyFactory(omics_processing=omics_processing1, name="assembly2")
-    fakes.MetagenomeAssemblyFactory(omics_processing=omics_processing2, name="assembly3")
+    fakes.MetagenomeAssemblyFactory(was_informed_by=[omics_processing1], name="assembly1")
+    fakes.MetagenomeAssemblyFactory(was_informed_by=[omics_processing1], name="assembly2")
+    fakes.MetagenomeAssemblyFactory(was_informed_by=[omics_processing2], name="assembly3")
 
-    fakes.MetagenomeAnnotationFactory(omics_processing=omics_processing1, name="annotation1")
-    fakes.MetagenomeAnnotationFactory(omics_processing=omics_processing1, name="annotation2")
-    fakes.MetagenomeAnnotationFactory(omics_processing=omics_processing2, name="annotation3")
+    fakes.MetagenomeAnnotationFactory(was_informed_by=[omics_processing1], name="annotation1")
+    fakes.MetagenomeAnnotationFactory(was_informed_by=[omics_processing1], name="annotation2")
+    fakes.MetagenomeAnnotationFactory(was_informed_by=[omics_processing2], name="annotation3")
 
-    fakes.MetaproteomicAnalysisFactory(omics_processing=omics_processing1, name="analysis1")
-    fakes.MetaproteomicAnalysisFactory(omics_processing=omics_processing1, name="analysis2")
-    fakes.MetaproteomicAnalysisFactory(omics_processing=omics_processing2, name="analysis3")
+    fakes.MetaproteomicAnalysisFactory(was_informed_by=[omics_processing1], name="analysis1")
+    fakes.MetaproteomicAnalysisFactory(was_informed_by=[omics_processing1], name="analysis2")
+    fakes.MetaproteomicAnalysisFactory(was_informed_by=[omics_processing2], name="analysis3")
 
-    fakes.MAGsAnalysisFactory(omics_processing=omics_processing1, name="mags1")
-    fakes.MAGsAnalysisFactory(omics_processing=omics_processing1, name="mags2")
-    fakes.MAGsAnalysisFactory(omics_processing=omics_processing2, name="mags3")
+    fakes.MAGsAnalysisFactory(was_informed_by=[omics_processing1], name="mags1")
+    fakes.MAGsAnalysisFactory(was_informed_by=[omics_processing1], name="mags2")
+    fakes.MAGsAnalysisFactory(was_informed_by=[omics_processing2], name="mags3")
 
-    fakes.ReadBasedAnalysisFactory(omics_processing=omics_processing1, name="reads1")
-    fakes.ReadBasedAnalysisFactory(omics_processing=omics_processing1, name="reads2")
-    fakes.ReadBasedAnalysisFactory(omics_processing=omics_processing2, name="reads3")
+    fakes.ReadBasedAnalysisFactory(was_informed_by=[omics_processing1], name="reads1")
+    fakes.ReadBasedAnalysisFactory(was_informed_by=[omics_processing1], name="reads2")
+    fakes.ReadBasedAnalysisFactory(was_informed_by=[omics_processing2], name="reads3")
 
-    fakes.NOMAnalysisFactory(omics_processing=omics_processing1, name="nom1")
-    fakes.NOMAnalysisFactory(omics_processing=omics_processing1, name="nom2")
-    fakes.NOMAnalysisFactory(omics_processing=omics_processing2, name="nom3")
+    fakes.NOMAnalysisFactory(was_informed_by=[omics_processing1], name="nom1")
+    fakes.NOMAnalysisFactory(was_informed_by=[omics_processing1], name="nom2")
+    fakes.NOMAnalysisFactory(was_informed_by=[omics_processing2], name="nom3")
 
-    fakes.MetabolomicsAnalysisFactory(omics_processing=omics_processing1, name="metab1")
-    fakes.MetabolomicsAnalysisFactory(omics_processing=omics_processing1, name="metab2")
-    fakes.MetabolomicsAnalysisFactory(omics_processing=omics_processing2, name="metab3")
+    fakes.MetabolomicsAnalysisFactory(was_informed_by=[omics_processing1], name="metab1")
+    fakes.MetabolomicsAnalysisFactory(was_informed_by=[omics_processing1], name="metab2")
+    fakes.MetabolomicsAnalysisFactory(was_informed_by=[omics_processing2], name="metab3")
     db.commit()
 
     # test omics_processing not associated with biosamples
