@@ -18,6 +18,7 @@ import AppBanner from '@/components/AppBanner.vue';
 import AttributeItem from '@/components/Presentation/AttributeItem.vue';
 import IndividualTitle from '@/views/IndividualResults/IndividualTitle.vue';
 import TeamInfo from '@/components/TeamInfo.vue';
+import gold from '@/assets/GOLD.png';
 /**
  * Override citations for certain DOIs
  */
@@ -28,6 +29,7 @@ const CitationOverrides: CitationOverridesType = {
   '10.46936/10.25585/60000017': 'Doktycz, M. (2020) BioScales - Defining plant gene function and its connection to ecosystem nitrogen and carbon cycling [Data set]. DOE Joint Genome Institute. https://doi.org/10.46936/10.25585/60000017',
 };
 const GoldStudyLinkBase = 'https://gold.jgi.doe.gov/study?id=';
+const BioprojectLinkBase = 'https://bioregistry.io/';
 
 export default defineComponent({
 
@@ -109,6 +111,15 @@ export default defineComponent({
         });
       }
       return links;
+    });
+
+    const bioprojectLinks = computed(() => {
+      if (item.value?.annotations?.insdc_bioproject_identifiers) {
+        return item.value.annotations.insdc_bioproject_identifiers.map((id) => (
+          BioprojectLinkBase + id
+        ));
+      }
+      return [];
     });
 
     function relatedTypeDescription(relatedType: string) {
@@ -206,6 +217,7 @@ export default defineComponent({
       CitationOverrides,
       GoldStudyLinkBase,
       goldLinks,
+      bioprojectLinks,
       data: dois,
       item,
       studyData,
@@ -219,14 +231,7 @@ export default defineComponent({
       fieldDisplayName,
       seeStudyInContext,
       parentStudies,
-      images: {
-        // eslint-disable-next-line global-require
-        gold: require('@/assets/GOLD.png'),
-        // eslint-disable-next-line global-require
-        ess: require('@/assets/ESS.png'),
-        // eslint-disable-next-line global-require
-        massive: require('@/assets/massive.png'),
-      },
+      gold,
     };
   },
 });
@@ -279,17 +284,20 @@ export default defineComponent({
             </v-list>
             <template
               v-if="
-                goldLinks.size > 0 ||
-                  (item.relevant_protocols && item.relevant_protocols.length > 0) ||
+                goldLinks.size > 0 || bioprojectLinks.length > 0 ||
+                  (item.protocol_link && item.protocol_link.length > 0) ||
                   item.principal_investigator_websites.length > 0"
             >
               <div class="display-1">
                 Additional Resources
               </div>
               <v-list
-                v-if="goldLinks.size > 0 || (item.relevant_protocols && item.relevant_protocols.length > 0) || item.principal_investigator_websites.length > 0"
+                v-if="
+                  goldLinks.size > 0 || bioprojectLinks.length > 0 ||
+                    (item.protocol_link && item.protocol_link.length > 0) ||
+                    item.principal_investigator_websites.length > 0"
               >
-                <v-list-item v-if="item.relevant_protocols">
+                <v-list-item v-if="item.protocol_link">
                   <v-list-item-avatar>
                     <v-icon>mdi-file-document</v-icon>
                   </v-list-item-avatar>
@@ -300,16 +308,16 @@ export default defineComponent({
                   </v-list-item-content>
                 </v-list-item>
                 <AttributeItem
-                  v-for="proto in (item.relevant_protocols || [])"
+                  v-for="proto in (item.protocol_link || [])"
                   :key="proto"
                   style="padding-left: 60px;"
                   v-bind="{
                     item,
                     link: { name: proto, target: proto},
-                    field: 'relevant_protocols' }
+                    field: 'protocol_link' }
                   "
                 />
-                <v-list-item v-if="goldLinks.size > 0 || item.principal_investigator_websites.length > 0">
+                <v-list-item v-if="goldLinks.size > 0 || bioprojectLinks.length > 0 || item.principal_investigator_websites.length > 0">
                   <v-list-item-avatar>
                     <v-icon>mdi-file-document</v-icon>
                   </v-list-item-avatar>
@@ -330,7 +338,19 @@ export default defineComponent({
                       target: link
                     }
                   }"
-                  :image="images.gold"
+                  :image="gold"
+                />
+                <AttributeItem
+                  v-for="link in bioprojectLinks"
+                  :key="link"
+                  style="padding-left: 60px;"
+                  v-bind="{
+                    item,
+                    link: {
+                      name: 'BioProject',
+                      target: link
+                    }
+                  }"
                 />
                 <template
                   v-for="site in (item.principal_investigator_websites || [])"
