@@ -1,14 +1,35 @@
 <script lang="ts">
-import { defineComponent, ref } from '@vue/composition-api';
-import Definitions from '@/definitions';
 import {
-  multiOmicsForm, multiOmicsFormValid, multiOmicsAssociations, templateChoiceDisabled, contextForm, canEditSubmissionMetadata,
+  defineComponent,
+  ref,
+  onMounted,
+  watch,
+  nextTick,
+  computed,
+} from '@vue/composition-api';
+
+import Definitions from '@/definitions';
+import doiProviderValues from '@/schema';
+import {
+  multiOmicsForm, multiOmicsFormValid, multiOmicsAssociations, checkJGITemplates, canEditSubmissionMetadata, addAwardDoi, removeAwardDoi,
+  templateHasData, checkDoiFormat,
 } from '../store';
+import { AwardTypes, HARMONIZER_TEMPLATES } from '@/views/SubmissionPortal/types';
+
 import SubmissionDocsLink from './SubmissionDocsLink.vue';
 import SubmissionPermissionBanner from './SubmissionPermissionBanner.vue';
+import DataTypes from './DataTypes.vue';
+import DoeFacility from './DoeFacility.vue';
+
+const OTHER = 'OTHER';
 
 export default defineComponent({
-  components: { SubmissionDocsLink, SubmissionPermissionBanner },
+  components: {
+    DataTypes,
+    DoeFacility,
+    SubmissionDocsLink,
+    SubmissionPermissionBanner,
+  },
   setup() {
     const formRef = ref();
 
@@ -16,17 +37,129 @@ export default defineComponent({
       formRef.value.validate();
     }
 
+    const projectAwardValidationRules = () => [(v: string | undefined) => {
+      const facilityChosen = multiOmicsForm.facilities.length > 0;
+      if (!facilityChosen) {
+        return true;
+      }
+      const awardChosen = v !== undefined;
+      if (!awardChosen) {
+        return 'If submitting to a use facility, this field is required.';
+      }
+      return true;
+    }];
+    const otherAwardValidationRules = () => [(v: string | undefined) => {
+      const facilityChosen = multiOmicsForm.facilities.length > 0;
+      const awardChosen = multiOmicsForm.award !== undefined;
+      if (!facilityChosen && !awardChosen) {
+        return true;
+      }
+      if (!awardChosen || multiOmicsForm.award !== OTHER) {
+        return true;
+      }
+      const inputEmpty = v === undefined || v.trim().length === 0;
+      if (inputEmpty) {
+        return 'Please enter the kind of project award.';
+      }
+      return true;
+    }];
+
+    const doiRequired = computed(() => (
+      (multiOmicsForm.dataGenerated && multiOmicsForm.facilityGenerated)
+      || (!multiOmicsForm.dataGenerated && !multiOmicsForm.unknownDoi)
+    ));
+
+    const doiValueRules = (i: number) => (
+      [
+        (value: string) => {
+          const valid = !!value || !doiRequired.value;
+          return valid || 'An award DOI is required if data was or is going to be generated by a DOE facility.';
+        },
+        (value: string) => {
+          const hasProvider = multiOmicsForm.awardDois && multiOmicsForm.awardDois[i] && multiOmicsForm.awardDois[i].provider;
+          let valid = true;
+          if (hasProvider) {
+            valid = !!value;
+          }
+          return valid || 'DOI must be provided if a provider is selected.';
+        },
+        (value: string) => {
+          const valid = !value || checkDoiFormat(value);
+          return valid || 'DOI must be in the correct format.';
+        },
+      ]
+    );
+
+    const doiProviderRules = (i: number) => (
+      [
+        (provider: string) => {
+          const hasValue = multiOmicsForm.awardDois && multiOmicsForm.awardDois[i] && multiOmicsForm.awardDois[i].value !== '';
+          let valid = true;
+          if (hasValue || doiRequired.value) {
+            valid = !!provider;
+          }
+          return valid || 'A provider must be selected.';
+        },
+      ]
+    );
+
+    const revalidate = () => {
+      nextTick(() => formRef.value.validate());
+    };
+
+    function resetFields(field: string) {
+      if (field === 'dataGenerated') {
+        multiOmicsForm.facilityGenerated = undefined;
+        multiOmicsForm.doe = undefined;
+      }
+      multiOmicsForm.award = undefined;
+      multiOmicsForm.otherAward = undefined;
+      multiOmicsForm.mgCompatible = undefined;
+      multiOmicsForm.mgInterleaved = undefined;
+      multiOmicsForm.omicsProcessingTypes = [];
+      multiOmicsForm.mtCompatible = undefined;
+      multiOmicsForm.mtInterleaved = undefined;
+      multiOmicsForm.facilities = [];
+      revalidate();
+    }
+
+    watch(
+      () => multiOmicsForm.award,
+      (award) => {
+        const awardTypes = Object.values(AwardTypes) as string[];
+        if (award && awardTypes.includes(award)) {
+          multiOmicsForm.otherAward = '';
+        }
+        revalidate();
+      },
+    );
+
+    onMounted(() => {
+      formRef.value.validate();
+    });
+
     return {
+      OTHER,
+      addAwardDoi,
+      removeAwardDoi,
+      projectAwardValidationRules,
+      otherAwardValidationRules,
+      doiValueRules,
+      doiProviderRules,
+      revalidate,
+      resetFields,
       formRef,
       multiOmicsForm,
       multiOmicsAssociations,
       multiOmicsFormValid,
       Definitions,
-      templateChoiceDisabled,
-      contextForm,
+      HARMONIZER_TEMPLATES,
+      doiProviderValues,
       /* functions */
       reValidate,
       canEditSubmissionMetadata,
+      checkJGITemplates,
+      templateHasData,
     };
   },
 });
@@ -51,187 +184,283 @@ export default defineComponent({
       style="max-width: 1000px;"
       :disabled="!canEditSubmissionMetadata()"
     >
-      <div v-if="contextForm.dataGenerated === true">
-        <v-text-field
-          v-model="multiOmicsForm.GOLDStudyId"
-          label="GOLD Study ID"
-          :hint="Definitions.studyGoldID"
-          persistent-hint
-          outlined
-          dense
+      <v-radio-group
+        v-model="multiOmicsForm.dataGenerated"
+        label="Have data already been generated for your study? *"
+        :rules="[v => (v === true || v === false) || 'This field is required']"
+        :disabled="checkJGITemplates() || templateHasData(HARMONIZER_TEMPLATES.emsl.sampleDataSlot)"
+        @change="resetFields('dataGenerated')"
+      >
+        <v-radio
+          label="No"
+          :value="false"
         />
-        <v-text-field
-          v-model="multiOmicsForm.NCBIBioProjectId"
-          label="NCBI BioProject Accession"
-          :hint="Definitions.studyNCBIBioProjectAccession"
-          persistent-hint
-          outlined
-          dense
+        <v-radio
+          label="Yes"
+          :value="true"
         />
-        <v-combobox
-          v-model="multiOmicsForm.alternativeNames"
-          label="Alternative Names / IDs"
-          :hint="Definitions.studyAlternativeNames"
-          persistent-hint
-          deletable-chips
-          multiple
-          outlined
-          chips
-          small-chips
-          dense
-          append-icon=""
+      </v-radio-group>
+      <v-radio-group
+        v-if="multiOmicsForm.dataGenerated"
+        v-model="multiOmicsForm.facilityGenerated"
+        label="Was data generated at a DOE user facility (JGI, EMSL)? *"
+        :rules="[v => (v === true || v === false) || 'This field is required']"
+        :disabled="checkJGITemplates() || templateHasData(HARMONIZER_TEMPLATES.emsl.sampleDataSlot)"
+        @change="resetFields('facilityGenerated')"
+      >
+        <v-radio
+          label="No"
+          :value="false"
+          hide-details
+          class="mb-2 mt-0"
+        />
+        <v-radio
+          label="Yes"
+          :value="true"
+          hide-details
+          class="mb-2 mt-0"
+        />
+      </v-radio-group>
+      <div
+        v-if="multiOmicsForm.dataGenerated === true && multiOmicsForm.facilityGenerated"
+      >
+        <DoeFacility
+          @revalidate="revalidate"
         />
       </div>
-      <div class="text-h4">
-        Data types *
-      </div>
-      <div class="text-body-2 grey--text text--darken-2 mb-4">
-        {{ Definitions.metadataTypes }}
+      <v-radio-group
+        v-if="multiOmicsForm.dataGenerated === false"
+        v-model="multiOmicsForm.doe"
+        label="Are you submitting samples to a DOE user facility (JGI, EMSL)? *"
+        :rules="[v => (v === true || v === false) || 'This field is required']"
+        :disabled="checkJGITemplates() || templateHasData(HARMONIZER_TEMPLATES.emsl.sampleDataSlot)"
+        @change="resetFields('doe')"
+      >
+        <v-radio
+          label="No"
+          :value="false"
+          class="mb-2 mt-0"
+        />
+        <v-radio
+          label="Yes"
+          :value="true"
+          class="mb-2 mt-0"
+        />
+      </v-radio-group>
+
+      <v-radio-group
+        v-if="multiOmicsForm.dataGenerated === false && multiOmicsForm.doe"
+        v-model="multiOmicsForm.award"
+        label="What kind of project have you been awarded? *"
+        :rules="projectAwardValidationRules()"
+        class="pb-5"
+      >
+        <div class="d-flex">
+          <v-radio
+            value="CSP"
+            label="CSP"
+          />
+          <div class="ml-1">
+            (<a
+              href="https://jgi.doe.gov/user-programs/program-info/csp-overview/csp-annual-call/"
+              target="_blank"
+              rel="noopener noreferrer"
+            >Community Science Program</a>)
+          </div>
+        </div>
+
+        <div class="d-flex">
+          <v-radio
+            value="BERSS"
+            label="BERSS"
+          />
+          <div class="ml-1">
+            (<a
+              href="https://jgi.doe.gov/user-programs/other-programs/"
+              target="_blank"
+              rel="noopener noreferrer"
+            >Biological and Environmental Research Support Science</a>)
+          </div>
+        </div>
+
+        <div class="d-flex">
+          <v-radio
+            value="BRCs"
+            label="BRCs"
+          />
+          <div class="ml-1">
+            (<a
+              href="https://jgi.doe.gov/user-programs/other-programs/"
+              target="_blank"
+              rel="noopener noreferrer"
+            >Bioenergy Research Centers</a>)
+          </div>
+        </div>
+
+        <div class="d-flex">
+          <v-radio
+            value="FICUS"
+            label="FICUS"
+          />
+          <div class="ml-1">
+            (<a
+              href="https://www.emsl.pnnl.gov/basic/ficus-program/1872"
+              target="_blank"
+              rel="noopener noreferrer"
+            >Facilities Integrating Collaborations for User Science</a>)
+          </div>
+        </div>
+
+        <div class="d-flex">
+          <v-radio
+            value="MONet"
+            label="MONet"
+          />
+          <div class="ml-1">
+            (<a
+              href="https://www.emsl.pnnl.gov/monet"
+              target="_blank"
+              rel="noopener noreferrer"
+            >Molecular Observation Network</a>)
+          </div>
+        </div>
+
+        <v-radio
+          :value="OTHER"
+        >
+          <template #label>
+            <span class="mr-4">Other</span>
+            <v-text-field
+              v-model="multiOmicsForm.otherAward"
+              class="pa-0 ma-0"
+              dense
+              hide-details="auto"
+              outlined
+              :rules="otherAwardValidationRules()"
+            />
+          </template>
+        </v-radio>
+      </v-radio-group>
+      <div
+        v-if="multiOmicsForm.dataGenerated === false && multiOmicsForm.doe"
+        class="pb-4"
+      >
+        <DoeFacility
+          @revalidate="revalidate"
+        />
       </div>
 
+      <DataTypes
+        v-if="multiOmicsForm.dataGenerated === false && multiOmicsForm.doe === false"
+        legend="Which datatypes may be generated later?"
+        :show-data-compatibility-questions="false"
+        @revalidate="revalidate"
+      />
+      <DataTypes
+        v-if="multiOmicsForm.facilityGenerated === false"
+        legend="Which datatypes were generated?"
+        :show-data-compatibility-questions="true"
+        @revalidate="revalidate"
+      />
+
       <v-alert
-        v-if="templateChoiceDisabled"
+        v-if="checkJGITemplates() || templateHasData(HARMONIZER_TEMPLATES.emsl.sampleDataSlot)"
         type="warning"
       >
         <p class="text-h5">
-          Data type choice disabled
+          Some choices may be disabled
         </p>
-        Data types cannot be changed when there are already metadata rows in step 6.  To change the template, return to step 6 and remove all data.
+        Data types with data present in step 5 cannot be disabled. To change these templates, return to step 5 and remove data from the tabs you wish to disable.
+        You may add unselected data types at any time.
       </v-alert>
 
-      <!-- JGI -->
-      <div v-if="contextForm.facilities.includes('JGI') || contextForm.facilityGenerated === true">
-        <div class="text-h6">
-          Joint Genome Institute (JGI)
+      <div
+        v-if="multiOmicsForm.facilities.includes('EMSL') || multiOmicsForm.facilities.includes('JGI')"
+        class="my-4"
+      >
+        <div
+          v-for="_, i in multiOmicsForm.awardDois"
+          :key="`awardDois${i}`"
+          class="d-flex"
+        >
+          <v-card class="flex-grow-1 pa-4 mb-4">
+            <v-row>
+              <v-col
+                cols="12"
+                sm="8"
+              >
+                <v-text-field
+                  v-if="multiOmicsForm.awardDois !== null"
+                  v-model="multiOmicsForm.awardDois[i].value"
+                  label="Award DOI *"
+                  :hint="Definitions.doi"
+                  :rules="doiValueRules(i)"
+                  persistent-hint
+                  validate-on-blur
+                  outlined
+                  dense
+                  @change="revalidate"
+                >
+                  <template #message="{ message }">
+                    <span v-html="message" />
+                  </template>
+                </v-text-field>
+              </v-col>
+              <v-col
+                cols="12"
+                sm="4"
+              >
+                <v-select
+                  v-if="multiOmicsForm.awardDois !== null"
+                  v-model="multiOmicsForm.awardDois[i].provider"
+                  label="Award DOI Provider *"
+                  :hint="Definitions.dataDoiProvider"
+                  :items="doiProviderValues"
+                  persistent-hint
+                  outlined
+                  dense
+                  clearable
+                  :rules="doiProviderRules(i)"
+                  @change="revalidate"
+                >
+                  <template #message="{ message }">
+                    <span v-html="message" />
+                  </template>
+                </v-select>
+              </v-col>
+            </v-row>
+          </v-card>
+          <v-btn
+            v-if="(multiOmicsForm.dataGenerated || multiOmicsForm.facilities.includes('EMSL') || multiOmicsForm.facilities.includes('JGI')) && multiOmicsForm.awardDois !== null"
+            icon
+            class="pb-2"
+            :disabled="!(multiOmicsForm.facilities.length < multiOmicsForm.awardDois.length)"
+            @click="removeAwardDoi(i)"
+          >
+            <v-icon>mdi-minus-circle</v-icon>
+          </v-btn>
         </div>
         <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Metagenome"
-          value="mg-jgi"
-          :disabled="templateChoiceDisabled"
-          hide-details
-        />
-        <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Metagenome (Long Read)"
-          value="mg-lr-jgi"
-          :disabled="templateChoiceDisabled"
-          hide-details
-        />
-        <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Metatranscriptome"
-          value="mt-jgi"
-          :disabled="templateChoiceDisabled"
-          hide-details
-        />
-        <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Metabolome"
-          value="mb-jgi"
-          disabled
-          hide-details
-        />
-        <v-text-field
-          v-if="multiOmicsForm.omicsProcessingTypes.some((v) => v.endsWith('jgi'))"
-          v-model="multiOmicsForm.JGIStudyId"
-          :rules="[
-            v => !!v || 'JGI Proposal ID/Study ID is required when processing was done at JGI' ,
-            v => /^\d{6}$/.test(v) || 'JGI Proposal ID/Study ID must be a 6 digit numerical value'
-          ]"
-          label="JGI Proposal ID/Study ID *"
-          hint="This is the 6 digit ID assigned to your JGI Proposal and is required when completing metadata for samples to be sent to JGI for sequencing."
+          v-if="!multiOmicsForm.dataGenerated && (multiOmicsForm.facilities.includes('EMSL') || multiOmicsForm.facilities.includes('JGI'))"
+          v-model="multiOmicsForm.unknownDoi"
+          class="mt-0"
+          :label="`I don't know my award DOI`"
+          :hint="Definitions.unknownDoi"
           persistent-hint
-          class="mt-4"
-          outlined
-          validate-on-blur
-          dense
-          :error-messages="multiOmicsForm.JGIStudyId ? undefined : ['JGI Proposal ID/Study ID is required when processing was done at JGI']"
+          @change="revalidate"
         />
       </div>
-
-      <!-- EMSL -->
-      <div v-if="contextForm.facilities.includes('EMSL') || contextForm.facilityGenerated === true">
-        <div class="text-h6 mt-4">
-          Environmental Molecular Science Laboratory (EMSL)
-        </div>
-        <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Metaproteome"
-          value="mp-emsl"
-          :disabled="templateChoiceDisabled"
-          hide-details
-        />
-        <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Metabolome"
-          value="mb-emsl"
-          :disabled="templateChoiceDisabled"
-          hide-details
-        />
-        <v-checkbox
-          v-model="multiOmicsForm.omicsProcessingTypes"
-          label="Natural Organic Matter (FT-ICR MS)"
-          value="nom-emsl"
-          :disabled="templateChoiceDisabled"
-          hide-details
-        />
-        <v-text-field
-          v-if="multiOmicsForm.omicsProcessingTypes.some((v) => v.endsWith('emsl'))"
-          v-model="multiOmicsForm.studyNumber"
-          :rules="[
-            v => !!v || 'EMSL Study Number is required when processing was done at EMSL',
-            v => /^\d{5}$/.test(v) || 'EMSL Study Number must be a 5 digit numerical value'
-          ]"
-          hint="EMSL Study Number is required when processing was done at EMSL"
-          persistent-hint
-          label="EMSL Proposal / Study Number *"
-          class="mt-4"
-          outlined
-          validate-on-blur
-          dense
-          :error-messages="multiOmicsForm.studyNumber ? undefined : ['EMSL Study Number is required when processing was done at EMSL']"
-        />
-      </div>
-      <!-- Other -->
-      <div class="text-h6 mt-4">
-        Other Non-DOE
-      </div>
-      <v-checkbox
-        v-model="multiOmicsForm.omicsProcessingTypes"
-        label="Metagenome"
-        value="mg"
-        :disabled="templateChoiceDisabled"
-        hide-details
-      />
-      <v-checkbox
-        v-model="multiOmicsForm.omicsProcessingTypes"
-        label="Metatranscriptome"
-        value="mt"
-        :disabled="templateChoiceDisabled"
-        hide-details
-      />
-      <v-checkbox
-        v-model="multiOmicsForm.omicsProcessingTypes"
-        label="Metaproteome"
-        value="mp"
-        :disabled="templateChoiceDisabled"
-        hide-details
-      />
-      <v-checkbox
-        v-model="multiOmicsForm.omicsProcessingTypes"
-        label="Metabolome"
-        value="mb"
-        :disabled="templateChoiceDisabled"
-        hide-details
-      />
-      <v-checkbox
-        v-model="multiOmicsForm.omicsProcessingTypes"
-        label="Natural Organic Matter (FT-ICR MS)"
-        value="nom"
-        :disabled="templateChoiceDisabled"
-        hide-details
-      />
+      <v-btn
+        v-if="multiOmicsForm.facilities.includes('EMSL') || multiOmicsForm.facilities.includes('JGI')"
+        class="mb-4"
+        depressed
+        @click="addAwardDoi"
+      >
+        <v-icon class="pr-1">
+          mdi-plus-circle
+        </v-icon>
+        Add Award DOI
+      </v-btn>
     </v-form>
     <strong>* indicates required field</strong>
     <div class="d-flex mt-5">
@@ -250,7 +479,7 @@ export default defineComponent({
         color="primary"
         depressed
         :disabled="(!multiOmicsFormValid)"
-        :to="{ name: 'Environment Package' }"
+        :to="{ name: 'Sample Environment' }"
       >
         Go to next step
         <v-icon class="pl-1">
