@@ -9,7 +9,7 @@ import { read, writeFile, utils } from 'xlsx';
 import { api } from '@/data/api';
 import useRequest from '@/use/useRequest';
 
-import { HarmonizerApi } from './harmonizerApi';
+import HarmonizerApi from './harmonizerApi';
 import {
   packageName,
   sampleData,
@@ -29,10 +29,14 @@ import {
   isTestSubmission,
 } from './store';
 import {
+  DATA_MG_INTERLEAVED,
+  DATA_MG,
+  DATA_MT,
+  DATA_MT_INTERLEAVED,
   HARMONIZER_TEMPLATES,
   EMSL,
   JGI_MG,
-  JGT_MT,
+  JGI_MT,
   JGI_MG_LR,
   SuggestionsMode,
 } from '@/views/SubmissionPortal/types';
@@ -88,18 +92,9 @@ const TYPE_FIELD = ANALYSIS_TYPE;
 const COMMON_COLUMNS = [SAMP_NAME, SOURCE_MAT_ID, ANALYSIS_TYPE];
 
 const ALWAYS_READ_ONLY_COLUMNS = [
-  'dna_seq_project',
-  'rna_seq_project',
-  'dna_samp_id',
-  'rna_samp_id',
-  'rna_seq_project_pi',
-  'dna_seq_project_pi',
-  'dna_project_contact',
-  'rna_project_contact',
-  'proposal_rna',
-  'proposal_dna',
-  'rna_seq_project_name',
-  'dna_seq_project_name',
+  'jgi_seq_project',
+  'jgi_samp_id',
+  'jgi_seq_project_name',
 ];
 
 export default defineComponent({
@@ -151,6 +146,7 @@ export default defineComponent({
       }
       harmonizerApi.loadData(activeTemplateData.value);
       harmonizerApi.setInvalidCells(invalidCells.value[activeTemplateKey.value] || {});
+      harmonizerApi.changeVisibility(columnVisibility.value);
     });
 
     const validationErrors = computed(() => {
@@ -201,8 +197,145 @@ export default defineComponent({
       }
     });
 
+    function rowIsVisibleForTemplate(row: Record<string, any>, templateKey: string) {
+      const environmentKeys = templateList.value.filter((t) => HARMONIZER_TEMPLATES[t].status === 'published');
+      if (environmentKeys.includes(templateKey)) {
+        return true;
+      }
+      const row_types = row[TYPE_FIELD];
+      if (!row_types) {
+        return false;
+      }
+      if (templateKey === EMSL) {
+        return row_types.includes('lipidomics')
+          || row_types.includes('metaproteomics')
+          || row_types.includes('metabolomics')
+          || row_types.includes('natural organic matter');
+      }
+      if (templateKey === JGI_MG) {
+        return row_types.includes('metagenomics');
+      }
+      if (templateKey === JGI_MG_LR) {
+        return row_types.includes('metagenomics_long_read');
+      }
+      if (templateKey === JGI_MT) {
+        return row_types.includes('metatranscriptomics');
+      }
+      if (templateKey === DATA_MG) {
+        return row_types.includes('metagenomics');
+      }
+      if (templateKey === DATA_MG_INTERLEAVED) {
+        return row_types.includes('metagenomics');
+      }
+      if (templateKey === DATA_MT) {
+        return row_types.includes('metatranscriptomics');
+      }
+      if (templateKey === DATA_MT_INTERLEAVED) {
+        return row_types.includes('metatranscriptomics');
+      }
+      return false;
+    }
+
     // DataHarmonizer is a bit loose in its definition of empty cells. They can be null or and empty string.
     const isNonEmpty = (val: any) => val !== null && val !== '';
+
+    function synchronizeTabData(templateKey: string) {
+      const environmentKeys = templateList.value.filter((t) => HARMONIZER_TEMPLATES[t].status === 'published');
+      if (environmentKeys.includes(templateKey)) {
+        return;
+      }
+      const nextData = { ...sampleData.value };
+      const templateSlot = HARMONIZER_TEMPLATES[templateKey].sampleDataSlot;
+
+      const environmentSlots = templateList.value
+        .filter((t) => HARMONIZER_TEMPLATES[t].status === 'published')
+        .map((t) => HARMONIZER_TEMPLATES[t].sampleDataSlot);
+
+      if (!templateSlot || !environmentSlots) {
+        return;
+      }
+
+      // ensure the necessary keys exist in the data object
+      environmentSlots.forEach((slot) => {
+        if (!nextData[slot as string]) {
+          nextData[slot as string] = [];
+        }
+      });
+
+      if (!nextData[templateSlot]) {
+        nextData[templateSlot] = [];
+      }
+
+      // add/update any rows from the environment tabs to the active tab if they apply and if
+      // they aren't there already.
+      environmentSlots.forEach((environmentSlot) => {
+        nextData[environmentSlot as string].forEach((row) => {
+          const rowId = row[SCHEMA_ID];
+
+          const existing = nextData[templateSlot] && nextData[templateSlot].find((r) => r[SCHEMA_ID] === rowId);
+          if (!existing && rowIsVisibleForTemplate(row, templateKey)) {
+            const newRow = {} as Record<string, any>;
+            COMMON_COLUMNS.forEach((col) => {
+              newRow[col] = row[col];
+            });
+            nextData[templateSlot].push(newRow);
+            //update validation status for the tab, if data changed it needs to be revalidated
+            tabsValidated.value[templateKey] = false;
+          }
+          if (existing) {
+            COMMON_COLUMNS.forEach((col) => {
+              existing[col] = row[col];
+            });
+            //update validation status for the tab, if data changed it needs to be revalidated
+            tabsValidated.value[templateKey] = false;
+          }
+        });
+      });
+      // remove any rows from the active tab if they were removed from the environment tabs
+      // or no longer apply to the active tab
+      if (nextData[templateSlot].length > 0) {
+        nextData[templateSlot] = nextData[templateSlot].filter((row) => {
+          if (!rowIsVisibleForTemplate(row, templateKey)) {
+            return false;
+          }
+          //update validation status for the tab, if data changed it needs to be revalidated
+          tabsValidated.value[templateKey] = false;
+          const rowId = row[SCHEMA_ID];
+          return environmentSlots.some((environmentSlot) => {
+            const environmentRow = nextData[environmentSlot as string].findIndex((r) => r[SCHEMA_ID] === rowId);
+            return environmentRow >= 0;
+          });
+        });
+      }
+      sampleData.value = nextData;
+    }
+
+    /**
+     * This should be called whenever rows are removed from the data harmonizer.
+     * It ensures that row deletion is cascaded to facility templates from the main
+     * environment templates
+     */
+    const syncAndMergeTabsForRemovedRows = async () => {
+      mergeSampleData(
+        activeTemplate.value.sampleDataSlot,
+        harmonizerApi.exportJson(),
+      );
+      // If there are any sampleDataSlots populated that somehow are missing from
+      // the template list, make sure those data are updated as well.
+      Object.keys(sampleData.value).forEach((key) => {
+        // Loop through keys in the sampleData for the submission. Each
+        // key maps to a template. We have to find that template.
+        const [templateKey, template] = Object.entries(HARMONIZER_TEMPLATES).find(([, template]) => (
+          template?.sampleDataSlot === key
+        )) || [undefined, undefined];
+        if (template && templateKey) {
+          // If we found the template, synchronize the data
+          // Make sure we carry the deletion through to the sampleData
+          // The current tab's data needs to be updated first, then synchronized
+          synchronizeTabData(templateKey);
+        }
+      });
+    };
 
     const onDataChange = async (changes: any[]) => {
       // If we're in live suggestion mode and the user can edit the metadata, add the changes to a batch. Once the user
@@ -214,32 +347,27 @@ export default defineComponent({
         changeBatch.push(...nonEmptyChanges);
         debouncedSuggestionRequest();
       }
+      // If any changes touched the sample name or analysis/data type columns on an environment
+      // tab, we need to synch those changes to non-active tabs
+      const templateOrderedAttrNames = harmonizerApi.getOrderedAttributeNames(activeTemplate.value.schemaClass || '');
+      const shouldSynchronizeTabs = !!changes.find((change) => {
+        const isRelevantColumn = templateOrderedAttrNames[change[1]] === SAMP_NAME || templateOrderedAttrNames[change[1]] === ANALYSIS_TYPE;
+        const isNonemptyChange = isNonEmpty(change[2]) || isNonEmpty(change[3]);
+        return isNonemptyChange && isRelevantColumn;
+      });
 
       hasChanged.value += 1;
-      const data = harmonizerApi.exportJson();
-      mergeSampleData(activeTemplate.value.sampleDataSlot, data);
+      if (shouldSynchronizeTabs) {
+        syncAndMergeTabsForRemovedRows();
+      } else {
+        const data = harmonizerApi.exportJson();
+        mergeSampleData(activeTemplate.value.sampleDataSlot, data);
+      }
       saveRecord(); // This is a background save that we intentionally don't wait for
       tabsValidated.value[activeTemplateKey.value] = false;
     };
 
     const { request: schemaRequest, loading: schemaLoading } = useRequest();
-    onMounted(async () => {
-      const [schema, goldEcosystemTree] = await schemaRequest(() => Promise.all([
-        api.getSubmissionSchema(),
-        api.getGoldEcosystemTree(),
-      ]));
-      const r = document.getElementById('harmonizer-root');
-      if (r && schema) {
-        await harmonizerApi.init(r, schema, activeTemplate.value.schemaClass, goldEcosystemTree);
-        await nextTick();
-        harmonizerApi.loadData(activeTemplateData.value);
-        harmonizerApi.addChangeHook(onDataChange);
-        metadataSuggestions.value = getPendingSuggestions(root.$route.params.id, activeTemplate.value.schemaClass!);
-        if (!canEditSampleMetadata()) {
-          harmonizerApi.setTableReadOnly();
-        }
-      }
-    });
 
     async function jumpTo({ row, column }: { row: number; column: number }) {
       harmonizerApi.jumpToRowCol(row, column);
@@ -345,100 +473,9 @@ export default defineComponent({
     const doSubmit = () => submitRequest(async () => {
       const data = await harmonizerApi.exportJson();
       mergeSampleData(activeTemplate.value.sampleDataSlot, data);
-      await submit(root.$route.params.id, submissionStatus.SubmittedPendingReview);
+      await submit(root.$route.params.id, 'SubmittedPendingReview');
       submitDialog.value = false;
     });
-
-    function rowIsVisibleForTemplate(row: Record<string, any>, templateKey: string) {
-      const environmentKeys = templateList.value.filter((t) => HARMONIZER_TEMPLATES[t].status === 'published');
-      if (environmentKeys.includes(templateKey)) {
-        return true;
-      }
-      const row_types = row[TYPE_FIELD];
-      if (!row_types) {
-        return false;
-      }
-      if (templateKey === EMSL) {
-        return row_types.includes('metaproteomics')
-          || row_types.includes('metabolomics')
-          || row_types.includes('natural organic matter');
-      }
-      if (templateKey === JGI_MG) {
-        return row_types.includes('metagenomics');
-      }
-      if (templateKey === JGI_MG_LR) {
-        return row_types.includes('metagenomics_long_read');
-      }
-      if (templateKey === JGT_MT) {
-        return row_types.includes('metatranscriptomics');
-      }
-      return false;
-    }
-
-    function synchronizeTabData(templateKey: string) {
-      const environmentKeys = templateList.value.filter((t) => HARMONIZER_TEMPLATES[t].status === 'published');
-      if (environmentKeys.includes(templateKey)) {
-        return;
-      }
-      const nextData = { ...sampleData.value };
-      const templateSlot = HARMONIZER_TEMPLATES[templateKey].sampleDataSlot;
-
-      const environmentSlots = templateList.value
-        .filter((t) => HARMONIZER_TEMPLATES[t].status === 'published')
-        .map((t) => HARMONIZER_TEMPLATES[t].sampleDataSlot);
-
-      if (!templateSlot || !environmentSlots) {
-        return;
-      }
-
-      // ensure the necessary keys exist in the data object
-      environmentSlots.forEach((slot) => {
-        if (!nextData[slot as string]) {
-          nextData[slot as string] = [];
-        }
-      });
-
-      if (!nextData[templateSlot]) {
-        nextData[templateSlot] = [];
-      }
-
-      // add/update any rows from the environment tabs to the active tab if they apply and if
-      // they aren't there already.
-      environmentSlots.forEach((environmentSlot) => {
-        nextData[environmentSlot as string].forEach((row) => {
-          const rowId = row[SCHEMA_ID];
-
-          const existing = nextData[templateSlot] && nextData[templateSlot].find((r) => r[SCHEMA_ID] === rowId);
-          if (!existing && rowIsVisibleForTemplate(row, templateKey)) {
-            const newRow = {} as Record<string, any>;
-            COMMON_COLUMNS.forEach((col) => {
-              newRow[col] = row[col];
-            });
-            nextData[templateSlot].push(newRow);
-          }
-          if (existing) {
-            COMMON_COLUMNS.forEach((col) => {
-              existing[col] = row[col];
-            });
-          }
-        });
-      });
-      // remove any rows from the active tab if they were removed from the environment tabs
-      // or no longer apply to the active tab
-      if (nextData[templateSlot].length > 0) {
-        nextData[templateSlot] = nextData[templateSlot].filter((row) => {
-          if (!rowIsVisibleForTemplate(row, templateKey)) {
-            return false;
-          }
-          const rowId = row[SCHEMA_ID];
-          return environmentSlots.some((environmentSlot) => {
-            const environmentRow = nextData[environmentSlot as string].findIndex((r) => r[SCHEMA_ID] === rowId);
-            return environmentRow >= 0;
-          });
-        });
-      }
-      sampleData.value = nextData;
-    }
 
     async function downloadSamples() {
       templateList.value.forEach((templateKey) => {
@@ -475,7 +512,14 @@ export default defineComponent({
           const template = Object.values(HARMONIZER_TEMPLATES).find((template) => (
             template.excelWorksheetName === name || template.displayName === name
           ));
-          if (!template || !template.sampleDataSlot || !template.schemaClass) {
+          const templateSelected = templateList.value.find((selectedTemplate) => {
+            const templateName = HARMONIZER_TEMPLATES[selectedTemplate].displayName || '';
+            return (
+              template?.displayName === templateName
+              || template?.excelWorksheetName === templateName
+            );
+          });
+          if (!template || !template.sampleDataSlot || !template.schemaClass || !templateSelected) {
             notImported.push(name);
             return;
           }
@@ -520,6 +564,19 @@ export default defineComponent({
       reader.readAsArrayBuffer(file);
     }
 
+    /**
+     * Set up the hands on table with appropriate event handlers
+     * for interactions with the Data Harmonizer.
+     */
+    function addHooks() {
+      harmonizerApi.addChangeHook(onDataChange);
+      harmonizerApi.addRowRemovedHook(async () => {
+        syncAndMergeTabsForRemovedRows();
+        hasChanged.value += 1;
+        saveRecord();
+      });
+    }
+
     async function changeTemplate(index: number) {
       if (!harmonizerApi.ready.value) {
         return;
@@ -536,12 +593,32 @@ export default defineComponent({
       // When changing templates we may need to populate the common columns
       // from the environment tabs
       synchronizeTabData(nextTemplateKey);
-
       activeTemplateKey.value = nextTemplateKey;
       activeTemplate.value = nextTemplate;
       harmonizerApi.useTemplate(nextTemplate.schemaClass);
-      harmonizerApi.addChangeHook(onDataChange);
+      addHooks();
     }
+
+    onMounted(async () => {
+      const [schema, goldEcosystemTree] = await schemaRequest(() => Promise.all([
+        api.getSubmissionSchema(),
+        api.getGoldEcosystemTree(),
+      ]));
+      const r = document.getElementById('harmonizer-root');
+      if (r && schema) {
+        await harmonizerApi.init(r, schema, activeTemplate.value.schemaClass, goldEcosystemTree);
+        await nextTick();
+        harmonizerApi.loadData(activeTemplateData.value);
+        addHooks();
+        metadataSuggestions.value = getPendingSuggestions(
+          root.$route.params.id,
+          activeTemplate.value.schemaClass!,
+        );
+        if (!canEditSampleMetadata()) {
+          harmonizerApi.setTableReadOnly();
+        }
+      }
+    });
 
     return {
       user,
@@ -939,7 +1016,7 @@ export default defineComponent({
       <v-btn
         color="gray"
         depressed
-        :to="{ name: 'Environment Package' }"
+        :to="{ name: 'Sample Environment' }"
       >
         <v-icon class="pr-1">
           mdi-arrow-left-circle
@@ -969,11 +1046,11 @@ export default defineComponent({
             <v-btn
               color="success"
               depressed
-              :disabled="!canSubmit || status !== submissionStatus.InProgress || submitCount > 0"
+              :disabled="!canSubmit || status !== 'InProgress' || submitCount > 0"
               :loading="submitLoading"
               @click="submitDialog = true"
             >
-              <span v-if="status === submissionStatus.SubmittedPendingReview || submitCount">
+              <span v-if="status === 'SubmittedPendingReview' || submitCount">
                 <v-icon>mdi-check-circle</v-icon>
                 Submitted
               </span>
